指摘事項を修正、クラス化の変更
