+++ data_gorvernance/library/utils/dg_web/api.py
@@ -10,184 +10,178 @@ import requests
 from requests.exceptions import RequestException
 
 from ..error import UnauthorizedError, NotFoundContentsError
+from library.utils.config import connect as con_config
 
+class Api():
 
-def get_govsheet_schema(scheme:str, domain:str)->dict:
-    """ ガバナンスシートのスキーマを取得する関数です。
+    def __init__(self) -> None:
+        self._dgweb_base_url = con_config.get('DG_WEB', 'scheme_domain')
 
-    Args:
-        scheme (str): スキームを設定します。
-        domain (str): ドメインを設定します。
+    def get_govsheet_schema(self, scheme_domain:str)->dict:
+        """ ガバナンスシートのスキーマを取得する関数です。
 
-    Returns:
-        dict: ガバナンスシートのスキーマを返す。
+        Args:
+            scheme_domain (str): スキーマを設定します。
 
-    """
-    sub_url = '/schemas/govSheet'
-    api_url = parse.urlunparse((scheme, domain, sub_url, "", "", ""))
-    response = requests.get(url=api_url)
-    response.raise_for_status()
-    return response.json()
+        Returns:
+            dict: ガバナンスシートのスキーマを返す。
 
+        """
+        sub_url = '/schemas/govSheet'
+        api_url = scheme_domain + sub_url
+        response = requests.get(url=api_url)
+        response.raise_for_status()
+        return response.json()
 
-def get_metadata_schema(scheme:str, domain:str)->dict:
-    """ メタデータのスキーマを取得する関数です。
+    def get_metadata_schema(self, scheme_domain:str)->dict:
+        """ メタデータのスキーマを取得する関数です。
 
-    Args:
-        scheme (str): スキームを設定します。
-        domain (str): ドメインを設定します。
+        Args:
+            scheme_domain (str): スキーマを設定します。
 
-    Returns:
-        dict: メタデータのスキーマを返す。
+        Returns:
+            dict: メタデータのスキーマを返す。
 
-    """
-    sub_url = '/schemas/metadata'
-    api_url = parse.urlunparse((scheme, domain, sub_url, "", "", ""))
-    response = requests.get(url=api_url)
-    response.raise_for_status()
-    return response.json()
+        """
+        sub_url = '/schemas/metadata'
+        api_url = scheme_domain + sub_url
+        response = requests.get(url=api_url)
+        response.raise_for_status()
+        return response.json()
 
+    def check_governedrun_token(self, token:str)->bool:
+        """ Governed Runのトークンの有効性を確認する関数です。
 
-def check_governedrun_token(scheme:str, domain:str, token:str)->bool:
-    """ Governed Runのトークンの有効性を確認する関数です。
+        Args:
+            scheme_domain (str): スキーマを設定します。
+            token (str): Governed Runのトークンを設定します。
 
-    Args:
-        scheme (str): スキームを設定します。
-        domain (str): ドメインを設定します。
-        token (str): Governed Runのトークンを設定します。
+        Returns:
+            bool: Governed Runのトークンが有効であればTrue、有効でなければFalseを返す。
 
-    Returns:
-        bool: Governed Runのトークンが有効であればTrue、有効でなければFalseを返す。
+        Raises:
+            RequestException: 通信エラー
 
-    Raises:
-        RequestException: 通信エラー
-
-    """
-    sub_url = '/checkToken'
-    api_url = parse.urlunparse((scheme, domain, sub_url, "", "", ""))
-    data = {
-        "grdmToken": None,
-        "govRunToken": token
-    }
-    response = requests.post(url=api_url, json=data)
-    if response.status_code == HTTPStatus.OK:
-        return True
-    elif response.status_code == HTTPStatus.UNAUTHORIZED:
+        """
+        sub_url = '/checkToken'
+        api_url = self._dgweb_base_url + sub_url
+        data = {
+            "grdmToken": None,
+            "govRunToken": token
+        }
+        response = requests.post(url=api_url, json=data)
+        if response.status_code == HTTPStatus.OK:
+            return True
+        elif response.status_code == HTTPStatus.UNAUTHORIZED:
+            return False
+        response.raise_for_status()
         return False
-    response.raise_for_status()
-    return False
 
+    def validate(self, scheme_domain:str, grdm_token:str, project_id:str, govrun_token:str=None, govsheet:dict=None, metadata:dict=None)->dict:
+        """ 検証する関数です。
 
-def validate(scheme:str, domain:str, grdm_token:str, project_id:str, govrun_token:str=None, govsheet:dict=None, metadata:dict=None)->dict:
-    """ 検証する関数です。
+        Args:
+            scheme_domain (str): スキーマを設定します。
+            grdm_token(str): GRDMのトークンを設定します。
+            project_id(str): プロジェクトidを設定します。
+            govrun_token(str|None): Governed Runのトークンを設定します。
+            govsheet(dict|None): ガバナンスシートを設定します。
+            metadata(dict|None): メアデータを設定します。
 
-    Args:
-        scheme(str): スキームを設定します。
-        domain(str): ドメインを設定します。
-        grdm_token(str): GRDMのトークンを設定します。
-        project_id(str): プロジェクトidを設定します。
-        govrun_token(str|None): Governed Runのトークンを設定します。
-        govsheet(dict|None): ガバナンスシートを設定します。
-        metadata(dict|None): メアデータを設定します。
+        Returns:
+            dict: 検証結果を返す。
 
-    Returns:
-        dict: 検証結果を返す。
+        Raises:
+            UnauthorizedError: 認証が通らない
+            RequestException: その他の通信エラー
 
-    Raises:
-        UnauthorizedError: 認証が通らない
-        RequestException: その他の通信エラー
+        """
+        sub_url = '/validations/submit'
+        api_url = scheme_domain + sub_url
+        data = {
+            "grdmToken": grdm_token,
+            "govRunToken": govrun_token,
+            "grdmProjectId": project_id,
+            "govSheet": govsheet,
+            "metadata": metadata
+        }
+        response = requests.post(url=api_url, json=data)
+        try:
+            response.raise_for_status()
+        except RequestException as e:
+            if response.status_code == HTTPStatus.UNAUTHORIZED:
+                raise UnauthorizedError(str(e)) from e
+            raise
+        return response.json()
 
-    """
-    sub_url = '/validations/submit'
-    api_url = parse.urlunparse((scheme, domain, sub_url, "", "", ""))
-    data = {
-        "grdmToken": grdm_token,
-        "govRunToken": govrun_token,
-        "grdmProjectId": project_id,
-        "govSheet": govsheet,
-        "metadata": metadata
-    }
-    response = requests.post(url=api_url, json=data)
-    try:
-        response.raise_for_status()
-    except RequestException as e:
-        if response.status_code == HTTPStatus.UNAUTHORIZED:
-            raise UnauthorizedError(str(e)) from e
-        raise
-    return response.json()
+    def get_validations(self, scheme_domain:str, grdm_token: str, project_id: str)->dict:
+        """ 検証結果を取得する関数です。
 
+        Args:
+            scheme_domain (str): スキーマを設定します。
+            grdm_token (str): GRDMのトークンを設定します。
+            project_id (str): プロジェクトidを設定します。
 
-def get_validations(scheme:str, domain:str, grdm_token: str, project_id: str)->dict:
-    """ 検証結果を取得する関数です。
+        Returns:
+            dict: 全ての検証結果を返す。
 
-    Args:
-        scheme (str): スキームを設定します。
-        domain (str): ドメインを設定します。
-        grdm_token (str): GRDMのトークンを設定します。
-        project_id (str): プロジェクトidを設定します。
+        Raises:
+            UnauthorizedError: 認証が通らない
+            NotFoundContentsError: 検証結果が存在しない
+            RequestException: その他の通信エラー
 
-    Returns:
-        dict: 全ての検証結果を返す。
+        """
+        sub_url = '/validations'
+        api_url = scheme_domain + sub_url
+        data = {
+            "grdmToken": grdm_token,
+            "grdmProjectId": project_id,
+        }
+        response = requests.post(url=api_url, json=data)
+        try:
+            response.raise_for_status()
+        except RequestException as e:
+            if response.status_code == HTTPStatus.UNAUTHORIZED:
+                raise UnauthorizedError(str(e)) from e
+            if response.status_code == HTTPStatus.NOT_FOUND:
+                raise NotFoundContentsError(str(e)) from e
+            raise
+        return response.json()
 
-    Raises:
-        UnauthorizedError: 認証が通らない
-        NotFoundContentsError: 検証結果が存在しない
-        RequestException: その他の通信エラー
+    def get_validations_validationId(self, scheme_domain:str, grdm_token: str, project_id: str, validation_id: str)->dict:
+        """ idを指定して検証結果を取得する関数です。
 
-    """
-    sub_url = '/validations'
-    api_url = parse.urlunparse((scheme, domain, sub_url, "", "", ""))
-    data = {
-        "grdmToken": grdm_token,
-        "grdmProjectId": project_id,
-    }
-    response = requests.post(url=api_url, json=data)
-    try:
-        response.raise_for_status()
-    except RequestException as e:
-        if response.status_code == HTTPStatus.UNAUTHORIZED:
-            raise UnauthorizedError(str(e)) from e
-        if response.status_code == HTTPStatus.NOT_FOUND:
-            raise NotFoundContentsError(str(e)) from e
-        raise
-    return response.json()
+        Args:
+            scheme_domain (str): スキーマを設定します。
+            grdm_token (str): GRDMのトークンを設定します。
+            project_id (str): プロジェクトidを設定します。
+            validation_id (str): 検証のidを設定します。
 
+        Returns:
+            dict: 指定したidの検証結果を返す。
 
-def get_validations_validationId(scheme:str, domain:str, grdm_token: str, project_id: str, validation_id: str)->dict:
-    """ idを指定して検証結果を取得する関数です。
+        Raises:
+            UnauthorizedError: 認証が通らない
+            NotFoundContentsError: 検証結果が存在しない
+            RequestException: その他の通信エラー
 
-    Args:
-        scheme (str): スキームを設定します。
-        domain (str): ドメインを設定します。
-        grdm_token (str): GRDMのトークンを設定します。
-        project_id (str): プロジェクトidを設定します。
-        validation_id (str): 検証のidを設定します。
-
-    Returns:
-        dict: 指定したidの検証結果を返す。
-
-    Raises:
-        UnauthorizedError: 認証が通らない
-        NotFoundContentsError: 検証結果が存在しない
-        RequestException: その他の通信エラー
-
-    """
-    sub_url = f'/validations/{validation_id}'
-    api_url = parse.urlunparse((scheme, domain, sub_url, "", "", ""))
-    data = {
-        "grdmToken": grdm_token,
-        "grdmProjectId": project_id,
-    }
-    params = {
-        "validationId": validation_id
-    }
-    response = requests.post(url=api_url, json=data, params=params)
-    try:
-        response.raise_for_status()
-    except RequestException as e:
-        if response.status_code == HTTPStatus.UNAUTHORIZED:
-            raise UnauthorizedError(str(e)) from e
-        if response.status_code == HTTPStatus.NOT_FOUND:
-            raise NotFoundContentsError(str(e)) from e
-        raise
-    return response.json()
+        """
+        sub_url = f'/validations/{validation_id}'
+        api_url = scheme_domain + sub_url
+        data = {
+            "grdmToken": grdm_token,
+            "grdmProjectId": project_id,
+        }
+        params = {
+            "validationId": validation_id
+        }
+        response = requests.post(url=api_url, json=data, params=params)
+        try:
+            response.raise_for_status()
+        except RequestException as e:
+            if response.status_code == HTTPStatus.UNAUTHORIZED:
+                raise UnauthorizedError(str(e)) from e
+            if response.status_code == HTTPStatus.NOT_FOUND:
+                raise NotFoundContentsError(str(e)) from e
+            raise
+        return response.json()
