+++ data_gorvernance/library/utils/dg_web/data.py
@@ -1,17 +1,18 @@
 """ ガバナンスシートやメタデータの設定値を扱うモジュールです。"""
-def need_govrun_token(govsheet:dict, metadata:dict)->bool:
-    """ Governed Runのトークンが必要かどうかを判定する関数です。
+class Data:
+    def need_govrun_token(self, govsheet:dict, metadata:dict)->bool:
+        """ Governed Runのトークンが必要かどうかを判定する関数です。
 
-    Args:
-        govsheet (dict): ガバナンスシートを設定します。
-        metadata (dict): メタデータを設定します。
+        Args:
+            govsheet (dict): ガバナンスシートを設定します。
+            metadata (dict): メタデータを設定します。
 
-    Returns:
-        bool: Governed Runのトークンが必要かを返す。
+        Returns:
+            bool: Governed Runのトークンが必要かを返す。
 
-    """
-    if govsheet.get("rerun", {}).get("rerunLevel") == "設定しない":
-        return False
-    if not metadata.get("runCrate"):
-        return False
-    return True
+        """
+        if govsheet.get("rerun", {}).get("rerunLevel") == "設定しない":
+            return False
+        if not metadata.get("runCrate"):
+            return False
+        return True
